SemanticQ_modular/requirements.txt
streamlit
langchain
langchain-groq
langchain-community
faiss-cpu
sentence-transformers
python-dotenv